requirements.txt
cycler==0.10.0
Cython==0.29.22
decorator==5.0.7
freetype-py==2.2.0
future==0.18.2
fvcore==0.1.5.post20210413
imageio==2.9.0
iopath==0.1.8
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.2
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
plyfile==0.7.3
portalocker==2.3.0
protobuf==3.15.8
pyglet==1.4.0b1
PyOpenGL==3.1.0
pyparsing==2.4.7
pyrender==0.1.32
python-dateutil==2.8.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
tabulate==0.8.9
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
trimesh==3.9.13
typing-extensions==3.7.4.3
yacs==0.1.8
